+++ typescript/sdk/src/token/config.ts
@@ -1,13 +1,3 @@
-import { z } from 'zod';
-
-import {
-  CollateralConfigSchema,
-  NativeConfigSchema,
-  SyntheticConfigSchema,
-  TokenMetadataSchema,
-  TokenRouterConfigSchema,
-} from './schemas.js';
-
 export enum TokenType {
   synthetic = 'synthetic',
   fastSynthetic = 'fastSynthetic',
@@ -33,17 +23,3 @@ export const gasOverhead = (tokenType: TokenType) => {
       return 68_000;
   }
 };
-
-export type TokenRouterConfig = z.infer<typeof TokenRouterConfigSchema>;
-export type NativeConfig = z.infer<typeof NativeConfigSchema>;
-export type CollateralConfig = z.infer<typeof CollateralConfigSchema>;
-
-function isCompliant<S extends Zod.Schema>(schema: S) {
-  return (config: unknown): config is z.infer<S> =>
-    schema.safeParse(config).success;
-}
-
-export const isSyntheticConfig = isCompliant(SyntheticConfigSchema);
-export const isCollateralConfig = isCompliant(CollateralConfigSchema);
-export const isNativeConfig = isCompliant(NativeConfigSchema);
-export const isTokenMetadata = isCompliant(TokenMetadataSchema);
