+++ typescript/sdk/src/token/schemas.test.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { constants, ethers } from 'ethers';
+import { ethers } from 'ethers';
 
 import { TokenType } from './config.js';
 import { WarpRouteDeployConfigSchema } from './schemas.js';
@@ -16,58 +16,43 @@ const NON_COLLATERAL_TYPES = [
   TokenType.synthetic,
   TokenType.syntheticUri,
   TokenType.fastSynthetic,
-  TokenType.native,
 ];
 
 describe('WarpRouteDeployConfigSchema refine', () => {
-  it('should require type address', () => {
-    const config: any = {
+  let config: any;
+  beforeEach(() => {
+    config = {
       arbitrum: {
         type: TokenType.collateral,
         token: SOME_ADDRESS,
+        owner: SOME_ADDRESS,
+        mailbox: SOME_ADDRESS,
       },
     };
+  });
+
+  it('should require token type', () => {
     expect(WarpRouteDeployConfigSchema.safeParse(config).success).to.be.true;
     delete config.arbitrum.type;
     expect(WarpRouteDeployConfigSchema.safeParse(config).success).to.be.false;
   });
 
   it('should require token address', () => {
-    const config: any = {
-      arbitrum: {
-        type: TokenType.collateral,
-        token: SOME_ADDRESS,
-      },
-    };
     expect(WarpRouteDeployConfigSchema.safeParse(config).success).to.be.true;
     delete config.arbitrum.token;
     expect(WarpRouteDeployConfigSchema.safeParse(config).success).to.be.false;
   });
 
-  it('should allow mailbox to be optional', () => {
-    const config: any = {
-      arbitrum: {
-        type: TokenType.collateral,
-        token: constants.AddressZero,
-        mailbox: SOME_ADDRESS,
-      },
-    };
+  it('should require mailbox address', () => {
     expect(WarpRouteDeployConfigSchema.safeParse(config).success).to.be.true;
     delete config.arbitrum.mailbox;
-    expect(WarpRouteDeployConfigSchema.safeParse(config).success).to.be.true;
+    expect(WarpRouteDeployConfigSchema.safeParse(config).success).to.be.false;
   });
 
   it('should throw if collateral type and token is empty', async () => {
     for (const type of COLLATERAL_TYPES) {
-      const config: any = {
-        arbitrum: {
-          type,
-          mailbox: SOME_ADDRESS,
-          name: 'Arby Coin',
-          symbol: 'ARBY',
-          totalSupply: '10000',
-        },
-      };
+      config.arbitrum.type = type;
+      config.arbitrum.token = undefined;
       expect(WarpRouteDeployConfigSchema.safeParse(config).success).to.be.false;
 
       // Set to some address
@@ -76,17 +61,23 @@ describe('WarpRouteDeployConfigSchema refine', () => {
     }
   });
 
-  it('should succeed if non-collateral type and token is empty', async () => {
+  it('should accept native type if token is empty', async () => {
+    config.arbitrum.type = TokenType.native;
+    config.arbitrum.token = undefined;
+    expect(WarpRouteDeployConfigSchema.safeParse(config).success).to.be.true;
+  });
+
+  it('should succeed if non-collateral type, token is empty, metadata is defined', async () => {
+    delete config.arbitrum.token;
+    config.arbitrum.totalSupply = '0';
+    config.arbitrum.name = 'name';
+
     for (const type of NON_COLLATERAL_TYPES) {
-      const config: any = {
-        arbitrum: {
-          type,
-          mailbox: SOME_ADDRESS,
-          name: 'Arby Coin',
-          symbol: 'ARBY',
-          totalSupply: '10000',
-        },
-      };
+      config.arbitrum.type = type;
+      config.arbitrum.symbol = undefined;
+      expect(WarpRouteDeployConfigSchema.safeParse(config).success).to.be.false;
+
+      config.arbitrum.symbol = 'symbol';
       expect(WarpRouteDeployConfigSchema.safeParse(config).success).to.be.true;
     }
   });
